apps/api/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
redis
rq

